src/requirements.txt
django==2.2.8
djangorestframework==3.10.3
pytz==2018.5
psycopg2==2.7.5
gunicorn==19.9.0
django-heroku==0.3.1
dj-database-url==0.5.0
django-cors-headers~=3.1.1
whitenoise==4.1.4 
cryptography~=2.8
drf-jwt~=1.13.3
PyJWT~=1.7.1
requests~=2.22.0
python-jose~=3.0
